Removed errors and some warnings for clang compiler

+++ src/devices/spy-server/spyserver-tcp-client.cpp
@@ -33,9 +33,9 @@
 #include <iostream>
 
 /*
- *	Thanks to Youssef Touil (well know from Airspy and spyServer),
- *	the structure of this program part is an adapted C++ translation
- *	of a C# fragment that he is using
+ *  Thanks to Youssef Touil (well know from Airspy and spyServer),
+ *  the structure of this program part is an adapted C++ translation
+ *  of a C# fragment that he is using
  */
 #ifdef __MINGW32__
 // #pragma comment(lib, "ws2_32.lib")
@@ -64,15 +64,15 @@ SpyServerTcpClient::SpyServerTcpClient(const QString & addr, i32 port,
   this->inBuffer = inBuffer;
   connected = false;
 
-#ifdef	__MINGW32__
+#ifdef  __MINGW32__
     WSAData wsaData;
     WSAStartup (MAKEWORD (2, 2), &wsaData);
     fprintf (stderr, "Client: Winsock DLL is %s\n",
                                      wsaData. szSystemStatus);
 #endif
   SendingSocket = socket(AF_INET, SOCK_STREAM, IPPROTO_TCP);
-#ifdef	__MINGW32__
-    if (SendingSocket == (i32)INVALID_SOCKET) {
+#ifdef  __MINGW32__
+    if (SendingSocket == INVALID_SOCKET) {
        fprintf (stderr, "Client: socket failed: Error code: %d\n", WSAGetLastError());
        WSACleanup ();
        return;
@@ -90,13 +90,13 @@ SpyServerTcpClient::SpyServerTcpClient(const QString & addr, i32 port,
                       (struct sockaddr *)&ServerAddr, sizeof(ServerAddr));
   if (RetCode != 0)
   {
-#ifdef	__MINGW32__
+#ifdef  __MINGW32__
        printf ("Client: connect() failed! Error code: %d\n", WSAGetLastError());
 #endif
 
     close(SendingSocket);
 
-#ifdef	__MINGW32__
+#ifdef  __MINGW32__
        WSACleanup();
 #endif
     return;
@@ -106,7 +106,7 @@ SpyServerTcpClient::SpyServerTcpClient(const QString & addr, i32 port,
     fprintf(stderr, "Client: connect() is OK, got connected...\n");
     fprintf(stderr, "Client: Ready for sending and/or receiving data...\n");
   }
-  start();	// this is the reader
+  start();  // this is the reader
   connected = true;
 }
 
@@ -174,10 +174,10 @@ void SpyServerTcpClient::run()
     else
     {
       sockaddr t;
-#ifndef	__MINGW32__
+#ifndef __MINGW32__
       u32 tt = 10;
 #else
-          i32	tt = 10;
+          i32   tt = 10;
 #endif
       unsigned long bytesAvailable = 0;
       i32 ret = ioctlsocket(SendingSocket, FIONREAD, &bytesAvailable);
